--- tests/bind-port.c
@@ -1,17 +1,0 @@
-#include <stdio.h>
-#include <winsock.h>
-
-int main()
-{
-    WSADATA wsa;
-    WSAStartup(MAKEWORD(2, 2), &wsa);
-
-    SOCKET s = socket(AF_INET, SOCK_STREAM, 0);
-
-    struct sockaddr_in addr = {};
-    addr.sin_family = AF_INET;
-    addr.sin_addr.s_addr = 0;
-    addr.sin_port = htons(0x29a);
-
-    bind(s, (struct sockaddr *) &addr, sizeof(addr));
-}
