--- tests/crash.c
@@ -1,7 +1,0 @@
-#include <stdio.h>
-
-int main()
-{
-    char *p = NULL;
-    *p = 0;
-}
