+++ tests/general/getcursorpos.c
@@ -1,0 +1,37 @@
+/*
+Cuckoo Sandbox - Automated Malware Analysis
+Copyright (C) 2010-2014 Cuckoo Sandbox Developers
+
+This program is free software: you can redistribute it and/or modify
+it under the terms of the GNU General Public License as published by
+the Free Software Foundation, either version 3 of the License, or
+(at your option) any later version.
+
+This program is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU General Public License for more details.
+
+You should have received a copy of the GNU General Public License
+along with this program.  If not, see <http://www.gnu.org/licenses/>.
+*/
+
+#include <stdio.h>
+#include <windows.h>
+
+int main(void)
+{
+    for (int i = 0; i < 10; i++) 
+	{
+        POINT p = {}; 
+		if ( GetCursorPos(&p) == 0 )
+		{
+			fprintf( stderr, "Error getcursorpos" );
+			return 1;
+		}
+        printf("{x: %ld, y: %ld}\n", p.x, p.y);
+        Sleep(100);
+    }
+
+	return 0;
+}
