--- tests/migrate-process.c
@@ -1,8 +1,0 @@
-#include <stdio.h>
-#include <windows.h>
-
-int main()
-{
-    // process identifier of explorer.exe on my VM
-    CloseHandle(OpenProcess(PROCESS_ALL_ACCESS, FALSE, 468));
-}
